requirements.txt
hydra-core
omegaconf
numpy